requirements.txt
alabaster==0.7.12
asgiref==3.3.4
astroid==2.5.6
Babel==2.9.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.2
django-debug-toolbar==2.2
docutils==0.17.1
idna==2.10
imagesize==1.2.0
isort==5.8.0
Jinja2==3.0.1
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==20.9
Pillow==8.2.0
Pygments==2.9.0
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.5
wrapt==1.12.1
